Preliminary "txn.discard()" version.
If "db_seq_t" and PY_LONG_LONG are not compatible, compiler should show
a warning while compiling, and the generated code would be incorrect but
safe to use. No crash. Added sanity check in the testunit to verify this
is not the case, and the datatypes are 64 bit width in fact.